requirements.txt
fastapi[all]
uvicorn[standard]
sqlAlchemy[all]
pyrebase4
pytest
python-dotenv
coverage
pg8000==1.30.3